BR change AGAIN ffs
Last change to BigReactors, well last change for 1.6.
Full stainless steel is little bit too much. So only few main parts now
requires that. Also lowered the fuel consumption to 4x.
If you don't like it, feel free to change it in the configs.

With this I am closing development for 1.6 InfiTech (will be fixing only
important bugs) and focusing on 1.7 version.